weight int --> double

+++ MyWorld.java
@@ -13,7 +13,7 @@ public class MyWorld extends World
      * Constructor for objects of class myWorld.
      * 
      */
-    public static int weight = 10;
+    public static double weight = 10.0;
     public MyWorld()
     {    
         // Create a new world with 600x400 cells with a cell size of 1x1 pixels.
